fix: 날짜 개산 수정

+++ src/main/java/com/fastcampus/reserve/interfaces/order/dto/response/OrderItemInfoResponse.java
@@ -10,7 +10,7 @@ public record OrderItemInfoResponse(
         String imageUrl,
         Integer maxGuestCount,
         Integer baseGuestCount,
-        Long day,
+        int day,
         String checkInTime,
         LocalDate checkInDate,
         String checkOutTime,
